refactor: move rubric tests into subfolder
